requirements.txt
-i https://pypi.org/simple
flask==2.2.2
line-bot-sdk==2.3.0
python-dotenv==0.21.0
transitions==0.9.0
graphviz==0.20.1
requests==2.28.1
bs4==0.0.1
